pyproject.toml
[tool.poetry]
name = "libermanov"
version = "0.1.0"
description = ""
authors = ["VasilevAlexandr97 <sashaorpix@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.0.0b3"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
